requirements.txt
duckdb==1.10.1
dbt-duckdb==1.10.1
polars==1.39.3
pyarrow==19.0.1
plotly==6.6.0
kaleido==0.2.1
matplotlib==3.10.8
seaborn==0.13.2
python-dotenv==1.0.1
pandas==2.2.3
